content/media/webaudio/test/test_bug856771.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 856771</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var context = new AudioContext();

  var source = context.createBufferSource();
  source.connect(context.destination);
  ok(true, "Nothing should leak");

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
